Initial commit with features to cleanup audio files after transcribing and downloading the video file by commenting/uncommetting lines.
