+++ Transcription_en.txt
@@ -1,0 +1,1 @@
+ Alright so here we are one of the elephants. The cool thing about these guys is that they have really, really, really long hunts. And that's cool. And that's pretty much all there is to say.
