+++ Transcription_es.txt
@@ -1,0 +1,1 @@
+ Hola, señorita, como te llamas, me amo Mike, me amo Mike, donde está el peño, feliz cumpleaños, que oras, que oras, la la la la, me gusta la biblioteca, vivo en la casa roja, yo tengo dos bicicletas, muchos clases así de nada, cuantos años tienes un momento, por favor, es el momento en el mes de los espelidos, el amor de los espelidos, mi mamá es bonita, mi gato es muy blanco, perdóname, perdóname, la la la la, uno dos tres y cuatro, cinco, seis, y teo, jo, no heve de ace, no remember how to say 11, Antonio banderas, nachos grande, y lleno de twist, es el momento en el mes de los espelidos, espelidos, amor de los espelidos,
