--- output_fr.txt
@@ -1,1 +1,0 @@
- Il n'est faction dans la République. Il y a les blocages et les blocages. Il faut pouvoir les lever quand ils empêchent la activité normale. Nous avons pu faire sur les ordres managers, par exemple, dans certaines grandes villes, ou sur les raffineries. Chaque fois qu'il faut le faire, ça a commencé. Ce sera fait pour qu'il n'est pas de blocage. Souvent, ce qu'on fait peut-être de gens. Je prends un engagement pour ce qui dépend de moi. Je pense qu'il faut qu'il y ait un chemin qui soit maintenant trouvé et qu'on se remette autour de la table.
